Fix Pay Resume when Reserve Court

+++ Sport4All/Sport4All/Model/Court.swift
@@ -37,7 +37,7 @@ struct Court: Codable {
 struct Price: Codable {
 	let id: Int
 	let price: Double
-	let time: Double
+	let time: Int
 	let updatedAt: Date?
 	let createdAt: Date?
